+++ packages/dashboard/src/pages/Wallets/components/Balance/index.tsx
@@ -6,13 +6,12 @@ import { Icons } from '@masknet/icons'
 import { FormattedCurrency, MiniNetworkSelector } from '@masknet/shared'
 import { DashboardRoutes } from '@masknet/shared-base'
 import { MaskColorVar } from '@masknet/theme'
-import { formatCurrency } from '@masknet/web3-shared-base'
+import { formatCurrency, getTokenUSDValue } from '@masknet/web3-shared-base'
 import { Box, Button, buttonClasses, styled, Typography } from '@mui/material'
 import { useDashboardI18N } from '../../../../locales/index.js'
 import { useIsMatched } from '../../hooks/index.js'
 import type { Web3Helper } from '@masknet/web3-helpers'
 import { Context } from '../../hooks/useContext.js'
-import { getTokenUSDValue } from '../../utils/getTokenUSDValue.js'
 
 const BalanceContainer = styled('div')(
     ({ theme }) => `
