+++ packages/dashboard/src/pages/Wallets/components/FungibleTokenTableRow/index.tsx
@@ -3,13 +3,12 @@ import { Box, Button, TableCell, TableRow, Tooltip, Typography } from '@mui/mate
 import { getMaskColor, makeStyles } from '@masknet/theme'
 import { FormattedCurrency, TokenIcon, WalletIcon } from '@masknet/shared'
 import { useChainContext, useNetworkDescriptors, useNetworkContext } from '@masknet/web3-hooks-base'
-import { CurrencyType, formatBalance, formatCurrency } from '@masknet/web3-shared-base'
+import { CurrencyType, formatBalance, formatCurrency, getTokenUSDValue } from '@masknet/web3-shared-base'
 import { NetworkPluginID } from '@masknet/shared-base'
 import { ChainId } from '@masknet/web3-shared-evm'
 import type { Web3Helper } from '@masknet/web3-helpers'
 import { useDashboardI18N } from '../../../../locales/index.js'
 import { ChangeNetworkTip } from './ChangeNetworkTip.js'
-import { getTokenUSDValue } from '../../utils/getTokenUSDValue.js'
 
 const useStyles = makeStyles()((theme) => ({
     icon: {
