--- packages/dashboard/src/pages/Wallets/utils/getTokenUSDValue.tsx
@@ -1,5 +1,0 @@
-import { CurrencyType } from '@masknet/web3-shared-base'
-
-export const getTokenUSDValue = (value?: {
-    [key in CurrencyType]?: string
-}) => (value ? Number.parseFloat(value[CurrencyType.USD] ?? '') : 0)
