+++ packages/web3-shared/base/src/utils/formatBalance.ts
