+++ packages/web3-shared/base/src/utils/formatCurrency.ts
