+++ packages/web3-shared/base/src/utils/formatElapsed.ts
@@ -1,4 +1,4 @@
-import { i18n } from '../../../shared-ui/locales_legacy/index.js'
+import { i18NextInstance } from '@masknet/shared-base'
 
 export function formatElapsed(from: number) {
     const msPerMinute = 60 * 1000
@@ -8,26 +8,26 @@ export function formatElapsed(from: number) {
     const msPerYear = msPerDay * 365
     const elapsed = Date.now() - from
     if (elapsed < msPerMinute)
-        return i18n.t('relative_time_seconds_ago', {
+        return i18NextInstance.t('relative_time_seconds_ago', {
             seconds: Math.round(elapsed / 1000),
         })
     if (elapsed < msPerHour)
-        return i18n.t('relative_time_minutes_ago', {
+        return i18NextInstance.t('relative_time_minutes_ago', {
             minutes: Math.round(elapsed / msPerMinute),
         })
     if (elapsed < msPerDay)
-        return i18n.t('relative_time_hours_ago', {
+        return i18NextInstance.t('relative_time_hours_ago', {
             hours: Math.round(elapsed / msPerHour),
         })
     if (elapsed < msPerMonth)
-        return i18n.t('relative_time_days_ago', {
+        return i18NextInstance.t('relative_time_days_ago', {
             days: Math.round(elapsed / msPerDay),
         })
     if (elapsed < msPerYear)
-        return i18n.t('relative_time_months_ago', {
+        return i18NextInstance.t('relative_time_months_ago', {
             months: Math.round(elapsed / msPerMonth),
         })
-    return i18n.t('relative_time_years_ago', {
+    return i18NextInstance.t('relative_time_years_ago', {
         years: Math.round(elapsed / msPerYear),
     })
 }
