+++ packages/web3-shared/base/src/utils/formatMarketCap.ts
