+++ packages/web3-shared/base/src/utils/formatPercentage.ts
