+++ packages/web3-shared/base/src/utils/formatSupply.ts
