+++ packages/web3-shared/base/src/utils/getTokenUSDValue.ts
@@ -1,0 +1,7 @@
+import { CurrencyType } from '@masknet/web3-shared-base'
+
+export function getTokenUSDValue(value?: {
+    [key in CurrencyType]?: string
+}) {
+    return value ? Number.parseFloat(value[CurrencyType.USD] ?? '') : 0
+}
