+++ packages/web3-shared/base/src/utils/index.ts
@@ -1,17 +1,20 @@
-export * from './address.js'
-export * from './balance.js'
+export * from './attemptUntil.js'
 export * from './constant.js'
-export * from './currency.js'
 export * from './fetch.js'
-export * from './function.js'
-export * from './locale.js'
-export * from './marketCap.js'
+export * from './formatBalance.js'
+export * from './formatCurrency.js'
+export * from './formatElapsed.js'
+export * from './formatMarketCap.js'
+export * from './formatPercentage.js'
+export * from './formatSupply.js'
+export * from './getLocale.js'
+export * from './getTokenUSDValue.js'
+export * from './isSameAddress.js'
+export * from './isValidTimestamp.js'
 export * from './number.js'
 export * from './pagination.js'
-export * from './percentage.js'
 export * from './resolver.js'
-export * from './supply.js'
-export * from './timestamp.js'
 export * from './token.js'
 export * from './types.js'
+
 export const HD_PATH_WITHOUT_INDEX_ETHEREUM = "m/44'/60'/0'/0"
