+++ packages/web3-shared/base/tests/marketCap.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { formatMarketCap } from '../src/utils/marketCap.js'
+import { formatMarketCap } from '../src/utils/formatMarketCap.js'
 
 describe('Market Cap format util test', () => {
     test.each([
